Break out the last few elaborator passes
