+++ hdl21/elab/elaborators/__init__.py
@@ -1,10 +1,13 @@
 """ 
-# Elaborator Classes Sub-Package
+# Elaborator Class(es)
+
+Recursive walkers of Hdl21 hierarchies which perform the mechanics of elaboration. 
 """
 
-from .elaborators import *
 from .slices import SliceResolver
 from .generators import GeneratorElaborator
 from .orphanage import Orphanage
 from .arrays import ArrayFlattener
 from .portrefs import ImplicitSignals, ImplicitBundles
+from .conntypes import SignalConnTypes, BundleConnTypes
+from .flatten_bundles import BundleFlattener
